scripts/requirements.txt
pandas
numpy
supabase
python-dotenv
tqdm
psycopg2-binary
boto3
